chore(ci): adjust path for windows in build.ps1

+++ scripts/build.ps1
@@ -9,15 +9,17 @@ cd $rootDir
 
 # Build the binary.
 if ($outFile -eq "") {
-  node "$rootDir/node_modules/.bin/caxa" --directory "$rootDir/build" --command "{{caxa}}/node_modules/.bin/node" "{{caxa}}/index.js" --output "./bin/vim-doge.exe"
+  node "$rootDir\node_modules\.bin\caxa" --directory "$rootDir/build" --command "{{caxa}}/node_modules/.bin/node" "{{caxa}}/index.js" --output "./bin/vim-doge.exe"
 } else {
-  node "$rootDir/node_modules/.bin/caxa" --directory "$rootDir/build" --command "{{caxa}}/node_modules/.bin/node" "{{caxa}}/index.js" --output "./bin/$outFile.exe"
+  node "$rootDir\node_modules\.bin\caxa" --directory "$rootDir/build" --command "{{caxa}}/node_modules/.bin/node" "{{caxa}}/index.js" --output "./bin/$outFile.exe"
 }
 
 # Archive the binary.
-$outFile = "$rootDir\bin\$outFile.zip"
-rm $rootDir\bin\*.zip
-echo "==> Archiving $rootDir\bin\vim-doge.exe -> $outFile"
-7z a -tzip "$outFile" "$rootDir\bin\vim-doge.exe"
+if ($outFile -ne "") {
+  $outFile = "$rootDir\bin\$outFile.zip"
+  rm $rootDir\bin\*.zip
+  echo "==> Archiving $rootDir\bin\vim-doge.exe -> $outFile"
+  7z a -tzip "$outFile" "$rootDir\bin\vim-doge.exe"
+}
 
 echo "Done building vim-doge binaries"
